+++ deployment/vc/svn.go
@@ -28,8 +28,8 @@ func (s *svnService) checkInfo(context *endly.Context, request *StatusRequest) (
 	extractRequest := exec.NewExtractRequest(target,
 		exec.DefaultOptions(),
 		exec.NewExtractCommand(fmt.Sprintf("svn info"), "", nil, nil,
-			endly.NewDataExtraction("origin", "^URL:[\\t\\s]+([^\\s]+)", false),
-			endly.NewDataExtraction("revision", "Revision:\\s+([^\\s]+)", false)),
+			endly.NewExtract("origin", "^URL:[\\t\\s]+([^\\s]+)", false),
+			endly.NewExtract("revision", "Revision:\\s+([^\\s]+)", false)),
 		exec.NewExtractCommand(fmt.Sprintf("svn stat"), "", nil, nil))
 
 	if err = endly.Run(context, extractRequest, runResponse); err != nil {