+++ runner/selenium/service_test.go
@@ -235,7 +235,7 @@ func TestSeleniumService_Calls(t *testing.T) {
 		Call: &runner.MethodCall{
 			Method:     "Click",
 			Parameters: []interface{}{},
-			Wait:       &endly.Repeatable{SleepTimeMs: 1},
+			Wait:       &endly.Repeater{SleepTimeMs: 1},
 		},
 		Selector: &runner.WebElementSelector{
 			By:    "css selector",
@@ -259,10 +259,10 @@ func TestSeleniumService_Calls(t *testing.T) {
 		Call: &runner.MethodCall{
 			Method:     "Text",
 			Parameters: []interface{}{},
-			Wait: &endly.Repeatable{
-				Repeat:       20,
-				SleepTimeMs:  100,
-				ExitCriteria: "$value:/WebDriver/",
+			Wait: &endly.Repeater{
+				Repeat:      20,
+				SleepTimeMs: 100,
+				Exit:        "$value:/WebDriver/",
 			},
 		},
 	})
@@ -326,10 +326,10 @@ func TestSeleniumService_Run(t *testing.T) {
 			{
 				Selector: runner.NewWebElementSelector("", "#output"),
 				Calls: []*runner.MethodCall{
-					runner.NewMethodCall("Text", &endly.Repeatable{
-						Repeat:       20,
-						SleepTimeMs:  100,
-						ExitCriteria: "$value:/WebDriver/",
+					runner.NewMethodCall("Text", &endly.Repeater{
+						Repeat:      20,
+						SleepTimeMs: 100,
+						Exit:        "$value:/WebDriver/",
 					}),
 				},
 			},